Rename test data types
